Terminada :)

+++ src/Mensajes/MensajeVenta.java
@@ -4,13 +4,13 @@ import General.Utilidades;
 
 public class MensajeVenta extends Mensaje {
     protected String nombreEmpresa;
-    protected int numTitulosAVender;
+    protected double numTitulosAVender;
 
 
     public MensajeVenta(int idOperacion, String nombreCliente, String dniCliente, String nombreEmpresa, TipoOperacion tipoOperacion) {//Constructor para ser llamado desde MensajeRespuestaVenta
         super(idOperacion, nombreCliente, dniCliente, nombreEmpresa, tipoOperacion);
     }
-    public MensajeVenta(int identificador, String nombreCliente,String dniCliente, String nombreEmpresa, int numTitulosAVender) {
+    public MensajeVenta(int identificador, String nombreCliente,String dniCliente, String nombreEmpresa, float numTitulosAVender) {
         super(identificador, nombreCliente, dniCliente);
         this.nombreCliente = nombreCliente;
         this.nombreEmpresa = nombreEmpresa;
@@ -21,12 +21,12 @@ public class MensajeVenta extends Mensaje {
         super(idOperacion,nombreCliente,dniCliente);
     }
 
-    public MensajeVenta(int idOperacion, String nombreCliente, String dniCliente, String nombreEmpresa, TipoOperacion tipoOperacion, int numTitulosAVender) {
+    public MensajeVenta(int idOperacion, String nombreCliente, String dniCliente, String nombreEmpresa, TipoOperacion tipoOperacion, double numTitulosAVender) {
         super(idOperacion, nombreCliente, dniCliente, nombreEmpresa, tipoOperacion);
         this.numTitulosAVender = numTitulosAVender;
     }
 
-    public int getNumTitulosAVender() {
+    public double getNumTitulosAVender() {
         return numTitulosAVender;
     }
